% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{add_notebook}
\alias{add_notebook}
\title{Add a notebook to an RO-Crate}
\usage{
add_notebook(rocrate, file_id, name = NULL, content = NULL)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{file_id}{Optional, notebook's filename.}

\item{name}{Optional, notebook's name.}

\item{content}{Optional, notebook's content.}
}
\description{
Add a notebook to an RO-Crate
}
