% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{add_readme}
\alias{add_readme}
\title{Add README file to an RO-Crate}
\usage{
add_readme(rocrate, text, filename = "README.md")
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{text}{Character vector with README content.}

\item{filename}{README filename.}
}
\value{
Updated rocrate object.
}
\description{
Add README file to an RO-Crate
}
