% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{add_software}
\alias{add_software}
\title{Add software application entity}
\usage{
add_software(rocrate, name, version = NULL)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{name}{Software name.}

\item{version}{Version string.}
}
\description{
Add software application entity
}
