% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{.validate_bagit_manifest}
\alias{.validate_bagit_manifest}
\title{Validate BagIt declaration}
\usage{
.validate_bagit_manifest(path, algo, manifest_suffix = "manifest")
}
\arguments{
\item{path}{String with full path to a compressed file contain an RO-Crate
bag, see \link[rocrateR]{bag_rocrate} for details. Alternatively, a path
to a directory containing an RO-Crate bag.}

\item{algo}{String with algorithm used to generate the RO-Crate bag
(default: \code{NULL}, which auto detects the algorithm from the
\verb{manifest-<algo>.txt} file, inside the bag). See \link[digest]{digest}
for more details on valid algorithms.}

\item{manifest_suffix}{String with suffix for the manifest file (default:
\code{"manifest"}).}
}
\value{
A list with \code{status} and \code{errors} identified.
}
\description{
Validate BagIt declaration
}
\keyword{internal}
