% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{.validate_bagit_payload_oxum}
\alias{.validate_bagit_payload_oxum}
\title{Validate BagIt Payload Oxum}
\usage{
.validate_bagit_payload_oxum(path)
}
\arguments{
\item{path}{String with full path to a compressed file contain an RO-Crate
bag, see \link[rocrateR]{bag_rocrate} for details. Alternatively, a path
to a directory containing an RO-Crate bag.}
}
\value{
A list with \code{status} and \code{errors} identified.
}
\description{
Validate BagIt Payload Oxum
}
\keyword{internal}
