% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{.capture_extra_entities}
\alias{.capture_extra_entities}
\title{Capture extra entities}
\usage{
.capture_extra_entities(...)
}
\arguments{
\item{...}{Optional entities to include in the RO-Crate (e.g., author).}
}
\value{
List with name additional entities.
}
\description{
Capture extra entities, given as inputs to \link{rocrate} and
\link{rocrate_5s}.
}
\keyword{internal}
