% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{.detect_manifest_algo}
\alias{.detect_manifest_algo}
\title{Detect BagIt archive's manifest algorith}
\usage{
.detect_manifest_algo(path)
}
\arguments{
\item{path}{Path to BagIt archive directory.}
}
\value{
String with the algorithm.
}
\description{
Detect BagIt archive's manifest algorith
}
\keyword{internal}
