% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{.find_bagit_root}
\alias{.find_bagit_root}
\title{Find BagIt root for an RO-Crate}
\usage{
.find_bagit_root(path)
}
\arguments{
\item{path}{String with path to RO-Crate bag.}
}
\value{
String with path to RO-Crate bag root (if any).
}
\description{
Find BagIt root for an RO-Crate
}
\keyword{internal}
