% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{.find_id_index}
\alias{.find_id_index}
\title{Find \verb{@id} index in RO-Crate}
\usage{
.find_id_index(rocrate, id)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{id}{String with the ID of the RO-Crate entity within \verb{@graph}}
}
\value{
Boolean vector with index for entity with \verb{@id}.
}
\description{
Find \verb{@id} index in RO-Crate. Useful to update a component of an entity in
the RO-Crate, add new component (e.g., author + corresponding \verb{@id}).
}
\keyword{internal}
