% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{.get_entity_ids}
\alias{.get_entity_ids}
\title{Get entity IDs for RO-Crate}
\usage{
.get_entity_ids(rocrate)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\value{
Vector entity IDs for \code{rocrate}.
}
\description{
Get entity IDs for RO-Crate
}
\keyword{internal}
