% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{.get_root_entity}
\alias{.get_root_entity}
\title{Get root entity for RO-Crate}
\usage{
.get_root_entity(rocrate)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\value{
Root entity for \code{rocrate}.
}
\description{
Get root entity for RO-Crate
}
\keyword{internal}
