% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{.validate_against_profile}
\alias{.validate_against_profile}
\title{Validate RO-Crate against profile}
\usage{
.validate_against_profile(rocrate, profile)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{profile}{List with profile requirements, see \code{R/profiles.R}}
}
\value{
Vector with errors (if any).
}
\description{
Validate RO-Crate against profile
}
\keyword{internal}
