% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{.validate_entity}
\alias{.validate_entity}
\alias{.validate_entity.character}
\alias{.validate_entity.list}
\alias{.validate_entity.numeric}
\title{Validate entity}
\usage{
.validate_entity(x, ..., ent_name = NULL, required = c("@id", "@type"))

\method{validate_entity}{character}(x, ..., ent_name = NULL, required = "type")

\method{validate_entity}{list}(x, ..., ent_name = NULL, required = c("@id", "@type"))

\method{validate_entity}{numeric}(x, ..., ent_name = NULL, required = "type")
}
\arguments{
\item{x}{New entity. If a single value (e.g., \code{character}, \code{numeric}) is
given, this is assumed to be the entity's \verb{@id}, if a \code{list} is given,
this is assumed to be a complete entity. Other options are objects of
type \code{person} and \code{organisation} (equivalenly \code{organization}).}

\item{...}{Optional additional entity values/properties. Used when \code{x} is
a single value.}

\item{ent_name}{String with the name of the entity.}

\item{required}{Vector with list of keys required for the entity to be valid.
(default: \code{c("@id", "@type")})}
}
\value{
Boolean value to indicate if the given entity is valid.
}
\description{
Validate entity
}
\keyword{internal}
