% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{.validate_entity_overview}
\alias{.validate_entity_overview}
\title{Entity validation overview}
\usage{
.validate_entity_overview(has_elements, required, ent_name = NULL)
}
\arguments{
\item{required}{Vector with list of keys required for the entity to be valid.
(default: \code{c("@id", "@type")})}

\item{ent_name}{String with the name of the entity.}
}
\value{
Boolean flag with result of entity validation
}
\description{
Entity validation overview
}
\keyword{internal}
