% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{.validate_json_syntax}
\alias{.validate_json_syntax}
\title{Validate JSON syntax for RO-Crate metadata file}
\usage{
.validate_json_syntax(metadata_path)
}
\arguments{
\item{metadata_path}{String with path to RO-Crate metadata file.}
}
\value{
Boolean indicating if \code{metadata_path} is valid.
}
\description{
Validate JSON syntax for RO-Crate metadata file
}
\keyword{internal}
