% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-json.R
\name{.validate_jsonld_compliance}
\alias{.validate_jsonld_compliance}
\title{Validate JSON-LD compliance of an RO-Crate}
\usage{
.validate_jsonld_compliance(rocrate)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\value{
Boolean value with compliance
}
\description{
Validate JSON-LD compliance of an RO-Crate
}
\keyword{internal}
