% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{.validate_rocrate}
\alias{.validate_rocrate}
\title{Validate RO-Crate and return list of errors identied}
\usage{
.validate_rocrate(rocrate, strict = FALSE)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{strict}{Boolean to indicate if stricter checks should be done (e.g.,
check profile specification).}
}
\value{
Vector of strings with errors identified
}
\description{
Validate RO-Crate and return list of errors identied
}
\keyword{internal}
