% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{.validate_rocrate_profile}
\alias{.validate_rocrate_profile}
\title{Validate RO-Crate profile}
\usage{
.validate_rocrate_profile(rocrate)
}
\arguments{
\item{rocrate}{A parsed RO-Crate object.}
}
\value{
Character vector of errors.
}
\description{
Performs profile-specific validation if \code{conformsTo} is declared.
}
\keyword{internal}
