% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{.validate_semantics}
\alias{.validate_semantics}
\title{Validate RO-Crate semantic structure}
\usage{
.validate_semantics(rocrate)
}
\arguments{
\item{rocrate}{A parsed RO-Crate object.}
}
\value{
Character vector of errors.
}
\description{
Performs semantic checks on the \verb{@graph}.
}
\keyword{internal}
