% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{.validate_structure}
\alias{.validate_structure}
\title{Validate minimal RO-Crate structure}
\usage{
.validate_structure(rocrate)
}
\arguments{
\item{rocrate}{A parsed RO-Crate object.}
}
\value{
Character vector of errors.
}
\description{
Ensures required top-level fields are present.
}
\keyword{internal}
