% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{entity}
\alias{entity}
\title{Create a data entity}
\usage{
entity(id, type, ...)
}
\arguments{
\item{id}{Scalar value with \verb{@id} for the entity (e.g., \code{character},
\code{numeric}).}

\item{type}{String with \verb{@type} for the entity (e.g., \code{Dataset}, \code{File}).}

\item{...}{Optional additional entity values/properties.}
}
\value{
List with an entity object.
}
\description{
Create a data entity
}
\examples{
# create entity for an organisation
organisation_uol <- rocrateR::entity(
  "https://ror.org/04xs57h96",
  type = "Organization",
  name = "University of Liverpool",
  url = "http://www.liv.ac.uk"
)

# create an entity for a person
person_rvd <- rocrateR::entity(
  "https://orcid.org/0000-0001-5036-8661",
  type = "Person",
  name = "Roberto Villegas-Diaz",
  affiliation = list(`@id` = organisation_uol$`@id`)
)
}
