% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{extract_content}
\alias{extract_content}
\title{Extract \code{File} entities content to files}
\usage{
extract_content(rocrate, path, overwrite = FALSE)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{path}{Directory where files will be written. RO-Crate root.}

\item{overwrite}{Logical. Overwrite existing files.}
}
\value{
Invisibly returns updated \code{rocrate} without contents.
}
\description{
Write the \code{content} field of \code{File} entities to disk using their \verb{@id}
as the filename.
}
