% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{is_rocrate}
\alias{is_rocrate}
\title{Check if object is an RO-Crate}
\usage{
is_rocrate(rocrate, strict = FALSE, error = TRUE)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{strict}{Boolean to indicate if stricter checks should be done (e.g.,
check profile specification).}

\item{error}{Boolean to indicate if the function should throw an error, if
any errors are found (default: \code{TRUE}).}
}
\value{
Boolean flag with RO-Crate validity.
}
\description{
Check if object is an RO-Crate
}
\examples{
basic_crate <- rocrateR::rocrate()

# check if the new crate is valid
basic_crate |>
  rocrateR::is_rocrate()
}
