% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{load_rocrate}
\alias{load_rocrate}
\alias{load_rocrate.rocrate}
\alias{load_rocrate.character}
\title{Load an RO-Crate from various input types}
\usage{
load_rocrate(x, ...)

\method{load_rocrate}{rocrate}(x, ..., strict = FALSE, verbose = FALSE)

\method{load_rocrate}{character}(x, ..., bagit_version = "1.0", strict = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A path (character) or an existing \link[rocrateR]{rocrate} object.}

\item{...}{Reserved for future extensions.}

\item{strict}{Logical. If TRUE, perform strict profile validation.}

\item{verbose}{Logical. If TRUE, emit diagnostic messages.}

\item{bagit_version}{String with version of BagIt used to generate the
RO-Crate bag (default: \code{"1.0"}).
See \doi{10.17487/RFC8493} for more details.}
}
\value{
A validated \code{rocrate} object.
}
\description{
High-level loader that can read:
\itemize{
\item A \code{ro-crate-metadata.json} file
\item A directory containing an RO-Crate
\item A BagIt-wrapped RO-Crate directory
\item A zipped BagIt RO-Crate archive
}
}
