% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{load_rocrate_bag}
\alias{load_rocrate_bag}
\title{Load an RO-Crate BagIt archive}
\usage{
load_rocrate_bag(path, algo = NULL, bagit_version = "1.0")
}
\arguments{
\item{path}{String with full path to a compressed file contain an RO-Crate
bag, see \link[rocrateR]{bag_rocrate} for details. Alternatively, a path
to a directory containing an RO-Crate bag.}

\item{algo}{String with algorithm used to generate the RO-Crate bag
(default: \code{NULL}, which auto detects the algorithm from the
\verb{manifest-<algo>.txt} file, inside the bag). See \link[digest]{digest}
for more details on valid algorithms.}

\item{bagit_version}{String with version of BagIt used to generate the
RO-Crate bag (default: \code{"1.0"}).
See \doi{10.17487/RFC8493} for more details.}
}
\value{
An object with the \link[rocrateR]{rocrate} class.
}
\description{
Load an RO-Crate BagIt archive
}
\examples{
# -------- SETUP --------
basic_crate <- rocrateR::rocrate()
# temp file
tmp_dir <- file.path(tempdir(), digest::digest(runif(1)))
tmp <- file.path(tmp_dir, "ro-crate-metadata.json")
dir.create(tmp_dir)

# bag RO-Crate
path_to_roc_bag <- rocrateR::bag_rocrate(basic_crate, path = tmp_dir)

# -------- INPUT: RO-Crate BagIt archive --------
rocrateR::load_rocrate_bag(path_to_roc_bag)

# -------- INPUT: Path --------
rocrateR::unbag_rocrate(path_to_roc_bag) |>
  rocrateR::load_rocrate_bag()

# delete temp directory
unlink(tmp_dir, recursive = TRUE)
}
\seealso{
Other RO-Crate BagIt archive functions: 
\code{\link{bag_rocrate}()},
\code{\link{is_rocrate_bag}()},
\code{\link{unbag_rocrate}()}
}
\concept{RO-Crate BagIt archive functions}
