% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{new_rocrate_validation}
\alias{new_rocrate_validation}
\title{Create a \link[rocrateR]{rocrate} validation result}
\usage{
new_rocrate_validation(
  errors = character(),
  warnings = character(),
  info = character()
)
}
\arguments{
\item{errors}{Character vector of errors.}

\item{warnings}{Character vector of warnings.}

\item{info}{Character vector of informational messages.}
}
\value{
A \code{rocrate_validation} object.
}
\description{
Internal constructor for validation results.
}
\keyword{internal}
