% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocrate.R
\name{rocrate}
\alias{rocrate}
\title{Create a new RO-Crate object}
\usage{
rocrate(
  ...,
  context = "https://w3id.org/ro/crate/1.2/context",
  conformsTo = gsub("\\\\/context$", "\\\\1", context)
)
}
\arguments{
\item{...}{Optional entities to include in the RO-Crate (e.g., author).}

\item{context}{String with URL to the version of the RO-Crate specification
to use. The context brings the defined terms into the metadata document.}

\item{conformsTo}{String with URL to the version of the RO-Crate
specification which this object conforms to. Conformance declares which
RO-Crate conventions of using those terms are being followed.}
}
\value{
RO-Crate object, list with an additional class, \code{rocrate}.
}
\description{
Create a new RO-Crate object. This object includes basic skeleton for the
RO-Crate metadata descriptor (\code{ro-crate-metadata.json}) file, as described
in the official documentation: https://w3id.org/ro/crate/1.2 >
\href{https://www.researchobject.org/ro-crate/specification/1.2/root-data-entity.html}{Root Data Entity}.
}
\examples{
rocrateR::rocrate()
}
