% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.entity}
\alias{summary.entity}
\title{Summary of an RO-Crate entity}
\usage{
\method{summary}{entity}(object, ...)
}
\arguments{
\item{object}{An entity object (list).}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class "summary.entity".
}
\description{
Provides a summary of a single RO-Crate entity.
}
