% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rocrate}
\alias{summary.rocrate}
\title{Summary of an RO-Crate}
\usage{
\method{summary}{rocrate}(object, ...)
}
\arguments{
\item{object}{An object of class "rocrate".}

\item{...}{Additional arguments (unused).}
}
\value{
An object of class "summary.rocrate".
}
\description{
Provides a summary of the contents of an RO-Crate, including the
number of entities, files and basic metadata information.
}
