% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{validate_rocrate}
\alias{validate_rocrate}
\title{Validate an RO-Crate}
\usage{
validate_rocrate(x, mode = c("stop", "report"), strict = FALSE)
}
\arguments{
\item{x}{A path (character) or an existing \link[rocrateR]{rocrate} object.}

\item{mode}{Either \code{"stop"} or \code{"report"}.}

\item{strict}{Logical. Enable profile validation.}
}
\value{
A \code{rocrate_validation} object (in report mode).
}
\description{
Performs structural, semantic and profile validation.
}
