% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{add_author}
\alias{add_author}
\title{Add an author to an RO-Crate}
\usage{
add_author(
  rocrate,
  name,
  orcid = NULL,
  affiliation = NULL,
  ror = NULL,
  set_author = TRUE
)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{name}{Author's name.}

\item{orcid}{Optional, ORCID identifier, for details see
\url{https://orcid.org}.}

\item{affiliation}{Optional, author's organisation.}

\item{ror}{Optional, ROR identifier for the affiliation, for details see
\url{https://ror.org}.}

\item{set_author}{Logical, used to indicate if the current user should be set
as the author of the RO-Crate.}
}
\value{
Updated RO-Crate object.
}
\description{
This helper creates an author entity and if \code{affiliation} is provided, then
creates an organisation entity for the user's affiliation.
}
