% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{add_entity}
\alias{add_entity}
\alias{add_entities}
\title{Add entity to RO-Crate}
\usage{
add_entity(rocrate, entity, overwrite = FALSE, verbose = FALSE)

add_entities(rocrate, entity, overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{entity}{Entity object (list) that contains at least the following
components: \verb{@id} and \verb{@type}. Alternatively, a list of entities.}

\item{overwrite}{Boolean flag to indicate if the entity (if found in the
given RO-Crate) should be overwritten.}

\item{verbose}{Boolean flag to indicate if status messages should be hidden
(default: \code{FALSE}).}
}
\value{
Updated RO-Crate with the new entity.
}
\description{
Add entity to RO-Crate
}
\examples{
basic_crate <- rocrateR::rocrate()

# create entity for an organisation
organisation_uol <- rocrateR::entity(
  "https://ror.org/04xs57h96",
  type = "Organization",
  name = "University of Liverpool",
  url = "http://www.liv.ac.uk"
)

# create an entity for a person
person_rvd <- rocrateR::entity(
  "https://orcid.org/0000-0001-5036-8661",
  type = "Person",
  name = "Roberto Villegas-Diaz",
  affiliation = list(`@id` = organisation_uol$`@id`)
)

basic_crate_v2 <- basic_crate |>
  rocrateR::add_entity(person_rvd) |>
  rocrateR::add_entity_value(
    id = "./",
    key = "author",
    value = list(`@id` = person_rvd$`@id`)
  ) |>
  rocrateR::add_entity(organisation_uol)
}
