% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{add_project}
\alias{add_project}
\title{Add project metadata}
\usage{
add_project(rocrate, name, description = NULL)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{name}{Project's name.}

\item{description}{Optional, project's description.}
}
\value{
Updated RO-Crate object.
}
\description{
Add project metadata
}
