% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{add_workflow}
\alias{add_workflow}
\title{Add a workflow to an RO-Crate}
\usage{
add_workflow(
  rocrate,
  file_id,
  name = NULL,
  description = NULL,
  language = "R",
  content = NULL
)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{file_id}{Filename of the workflow script.}

\item{name}{Workflow name.}

\item{description}{Optional description.}

\item{language}{Programming language (default \code{"R"}).}

\item{content}{Optional script contents.}
}
\value{
Updated RO-Crate object.
}
\description{
Register a workflow script (e.g. R, Python, Nextflow) as a
\code{ComputationalWorkflow} entity inside the RO-Crate.
}
