% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{.validate_bagit_declaration}
\alias{.validate_bagit_declaration}
\title{Validate BagIt declaration}
\usage{
.validate_bagit_declaration(path, algo = "sha512", bagit_version = "1.0")
}
\arguments{
\item{path}{String with full path to a compressed file contain an RO-Crate
bag, see \link[rocrateR]{bag_rocrate} for details. Alternatively, a path
to a directory containing an RO-Crate bag.}

\item{algo}{String with algorithm used to generate the RO-Crate bag
(default: \code{"sha512"}). See \link[digest]{digest} for more details.}

\item{bagit_version}{String with version of BagIt used to generate the
RO-Crate bag (default: \code{"1.0"}).
See \url{https://www.rfc-editor.org/info/rfc8493} for more details.}
}
\value{
A list with \code{status} and \code{errors} identified.
}
\description{
Validate BagIt declaration
}
\keyword{internal}
