% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{crate_project}
\alias{crate_project}
\title{Automatically create an RO-Crate from a project directory}
\usage{
crate_project(path, rocrate = rocrateR::rocrate())
}
\arguments{
\item{path}{String with project directory.}

\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\description{
Automatically create an RO-Crate from a project directory
}
