% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{crate_project}
\alias{crate_project}
\title{Automatically create an RO-Crate from a project directory}
\usage{
crate_project(path = NULL)
}
\arguments{
\item{path}{String with project directory.}
}
\description{
Automatically create an RO-Crate from a project directory
}
