% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{.is_valid_url}
\alias{.is_valid_url}
\title{Validate URL}
\source{
https://stackoverflow.com/a/73952264
}
\usage{
.is_valid_url(x, suffix = "")
}
\arguments{
\item{x}{String with URL to validate.}

\item{suffix}{String with any additional characters to match when validating
the given URL, \code{x}.}
}
\value{
Boolean value indicating if the given string (`x) is a valid URL.
}
\description{
Validate URL
}
\examples{
url <- c(
  "w3id.org/ro/crate/1.2/context",
  "http://w3id.org/ro/crate/1.2",
  "http://w3id.org/ro/crate/1.2/context",
  "https://w3id.org/ro/crate/1.2/context",
  "123",
  "https://w3id.org/ro/crate/1.1/context",
  "https://w3id.org/ro/crate/1.0/context"
)
rocrateR:::.is_valid_url(url)
rocrateR:::.is_valid_url(url, suffix = "/context")
}
\keyword{internal}
