% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{is_rocrate}
\alias{is_rocrate}
\title{Check if a given has the \code{rocrate} class}
\usage{
is_rocrate(rocrate)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\value{
Nothing, call for its side effect.
}
\description{
Check if a given has the \code{rocrate} class
}
\keyword{internal}
