% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{is_rocrate}
\alias{is_rocrate}
\title{Check if object is an RO-Crate}
\usage{
is_rocrate(rocrate)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\value{
Returns invisibly the input RO-Crate object.
}
\description{
Check if object is an RO-Crate
}
\examples{
basic_crate <- rocrateR::rocrate()

# check if the new crate is valid
basic_crate |>
  rocrateR::is_rocrate()
}
