% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rocrate.R
\name{load_rocrate}
\alias{load_rocrate}
\alias{load_rocrate.rocrate}
\alias{load_rocrate.character}
\title{Load an RO-Crate from various input types}
\usage{
load_rocrate(x, ...)

\method{load_rocrate}{rocrate}(x, ..., verbose = FALSE)

\method{load_rocrate}{character}(
  x,
  ...,
  verbose = FALSE,
  bagit_version = "1.0",
  load_content = FALSE,
  max_file_size = 10 * 1024^2
)
}
\arguments{
\item{x}{A path (character) or an existing \link[rocrateR]{rocrate} object.}

\item{...}{Reserved for future extensions.}

\item{verbose}{Logical. If \code{TRUE}, emit diagnostic messages.}

\item{bagit_version}{String with version of BagIt used to generate the
RO-Crate bag (default: \code{"1.0"}).
See \doi{10.17487/RFC8493} for more details.}

\item{load_content}{Logical. If \code{TRUE} , attempt to load external file
contents into the \code{content} field for entities of type \code{File}.}

\item{max_file_size}{Maximum file size (bytes) allowed when loading
content. Default 10MB.}
}
\value{
An \code{rocrate} object.
}
\description{
High-level loader that can read:
\itemize{
\item A \code{ro-crate-metadata.json} file
\item A directory containing an RO-Crate
\item A BagIt-wrapped RO-Crate directory
\item A zipped BagIt RO-Crate archive
}
}
\examples{
# -------- SETUP --------
basic_crate <- rocrateR::rocrate()
# temp file
tmp_dir <- file.path(tempdir(), digest::digest(basename(tempfile())))
tmp <- file.path(tmp_dir, "ro-crate-metadata.json")
dir.create(tmp_dir)

# -------- INPUT: RO-Crate --------
rocrateR::load_rocrate(basic_crate, verbose = TRUE)

# -------- INPUT: Path --------
# save RO-Crate
rocrateR::write_rocrate(basic_crate, path = tmp)

# load RO-Crate
## with file name
rocrateR::load_rocrate(tmp, verbose = TRUE)

## with directory
rocrateR::load_rocrate(tmp_dir, verbose = TRUE)

# delete temp directory
unlink(tmp_dir, recursive = TRUE)
}
