% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rocrate}
\alias{print.rocrate}
\title{Print RO-Crate}
\usage{
\method{print}{rocrate}(x, ...)
}
\arguments{
\item{x}{RO-Crate object, see \link{rocrate}.}

\item{...}{Optional arguments, not used.}
}
\value{
Invisibly the input RO-Crate, \code{x}.
}
\description{
Print RO-Crate, S3 method for class 'rocrate'. Creates a temporal JSON file,
which then is displayed with the \link{message} function.
}
\examples{
rocrateR::rocrate()
}
