% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocrateR-package.R
\docType{package}
\name{rocrateR-package}
\alias{rocrateR}
\alias{rocrateR-package}
\title{rocrateR: RO-Crate R Package Wrapper}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package for creating, manipulating and reading RO-Crates. Latest supported version of the specification: \url{https://w3id.org/ro/crate/1.2/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ResearchObject/ro-crate-r/}
  \item Report bugs at \url{https://github.com/ResearchObject/ro-crate-r/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@outlook.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Rebecca Wilson (\href{https://orcid.org/0000-0003-2294-593X}{ORCID})
  \item Olly Butters (\href{https://orcid.org/0000-0003-0354-8461}{ORCID})
  \item Stuart Wheater (\href{https://orcid.org/0009-0003-2419-1964}{ORCID})
}

Other contributors:
\itemize{
  \item Research Object community [copyright holder]
}

}
\keyword{internal}
