% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{unbag_rocrate}
\alias{unbag_rocrate}
\title{'Unbag' (extract) RO-Crate packed with BagIt}
\usage{
unbag_rocrate(path, output = path, quiet = FALSE)
}
\arguments{
\item{path}{String with path to compressed file containing an RO-Crate bag.}

\item{output}{String with target path where the contents will be extracted
(default: \code{path} - same path as input).}

\item{quiet}{Boolean flag to indicate if messages should be supressed
(default: \code{FALSE} - display messages).}
}
\value{
String with path to root of the RO-Crate, invisibly.
}
\description{
'Unbag' (extract) RO-Crate packed with BagIt
}
\seealso{
Other bag_rocrate: 
\code{\link{bag_rocrate}()},
\code{\link{is_rocrate_bag}()}
}
\concept{bag_rocrate}
