% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bag.R
\name{unbag_rocrate}
\alias{unbag_rocrate}
\title{'Unbag' (extract) RO-Crate packed with BagIt}
\usage{
unbag_rocrate(path, output = dirname(path), quiet = FALSE)
}
\arguments{
\item{path}{String with path to compressed file containing an RO-Crate bag.}

\item{output}{String with target path where the contents will be extracted
(default: \code{dirname(path)} - same directory as input \code{path}).}

\item{quiet}{Boolean flag to indicate if messages should be suppressed
(default: \code{FALSE} - display messages).}
}
\value{
String with path to root of the RO-Crate.
}
\description{
'Unbag' (extract) RO-Crate packed with BagIt
}
\examples{
# -------- SETUP --------
basic_crate <- rocrateR::rocrate()
# temp file
tmp_dir <- file.path(tempdir(), digest::digest(runif(1)))
tmp <- file.path(tmp_dir, "ro-crate-metadata.json")
dir.create(tmp_dir)

# bag RO-Crate
path_to_roc_bag <- rocrateR::bag_rocrate(basic_crate, path = tmp_dir)

# -------- INPUT: Path --------
rocrateR::unbag_rocrate(path_to_roc_bag)

# delete temp directory
unlink(tmp_dir, recursive = TRUE)
}
\seealso{
Other RO-Crate BagIt archive functions: 
\code{\link{bag_rocrate}()},
\code{\link{is_rocrate_bag}()},
\code{\link{load_rocrate_bag}()}
}
\concept{RO-Crate BagIt archive functions}
